registration 也切兩個
