+++ client/webpack.hot.config.js
@@ -9,7 +9,7 @@ const config = require('./webpack.base.config.js')
 config.entry.push(
   'webpack-dev-server/client?http://localhost:3500',
   'react-hot-loader/patch',
-  './app/bundles/OfficialSite/startup/registration',
+  './app/bundles/OfficialSite/startup/registrationHot',
   './app/assets/styles/OfficialSite/index.scss'
 )
